+++ src/App.js
@@ -2,6 +2,7 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Header from './Components/Header';
 import Home from './Components/Home';
 import About from './Components/About';
+import Details from './Components/Details';
 
 function App() {
   return (
@@ -9,7 +10,8 @@ function App() {
       <Header />
       <Routes>
         <Route path="/" index element={<Home />} />
-        <Route path="/about" index element={<About />} />
+        <Route path="/about" element={<About />} />
+        <Route path="/country/:countryName" element={<Details />} />
       </Routes>
     </Router>
   );
