+++ src/Components/Countries.jsx
@@ -1,9 +1,10 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
+import { NavLink } from 'react-router-dom';
 import { getCountries } from '../Redux/Countries/actions';
 import './Sass/Countries.scss';
 
-function Countries() {
+const Countries = () => {
   const dispatch = useDispatch();
   const countries = useSelector((state) => state.countries);
   useEffect(() => {
@@ -19,17 +20,16 @@ function Countries() {
           </div>
           <img src={country.flag} alt="covid-images" />
           <div className="flex">
-            <h3>Population:</h3>
-            <h3 className="t_bg">{country.population}</h3>
-          </div>
-          <div className="flex">
             <h3>Cases:</h3>
             <h3 className="t_bg">{country.cases}</h3>
+          </div>
+          <div className="flex">
+            <NavLink to={`/country/${country.country}`}>Details</NavLink>
           </div>
         </div>
       ))}
     </div>
   );
-}
+};
 
 export default Countries;
