+++ src/Components/Home.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import Continents from './Continents';
-// import Countries from './Countries';
+// import Continents from './Continents';
+import Countries from './Countries';
 import './Sass/Styles.scss';
 
 function Home() {
   return (
     <div className="container">
-      <Continents />
-      {/* <Countries /> */}
+      {/* <Continents /> */}
+      <Countries />
     </div>
   );
 }
